Registrar accounts: rotate is POST on the credentials route, not a sub-route
A `credentials` path component is .gitignored in every Aplisay repository,
deliberately, for literal credentials; #288 weakened that rule to commit a
route file under api/paths/phone-endpoints/{identifier}/credentials/, which
is the wrong fix. The negation goes, the directory goes, and rotation
becomes POST on the existing credentials.js route (GET reveals), with the
handler, its apiDoc, the test and the docs moved accordingly.
