+++ docs/phone-endpoints-api.md
@@ -214,7 +214,7 @@ A phone registration normally has the platform register *out* to the customer's 
 - `registrar` is `REGSERVER_REGISTRAR` on the deployment (503 code `registrar_unavailable` when it is unset), and is also the digest realm. `username` is unique across every organisation, since one realm serves them all.
 - `GET /api/phone-endpoints/{id}` returns `mode`, `kind`, and for a registrar account `bindings` (the PBX sockets the owning node last mirrored onto the row: `contact`, `received`, `transport`, `userAgent`, `registeredAt`, `expiresAt`, `node`) and `bindingsUpdatedAt`. Never the password.
 - `PUT` refuses `registrar`, `username`, `password` and `b2buaId` on a registrar account (400 code `registrar_identity_immutable`) and refuses a change of `mode` on any registration (400 code `mode_immutable`).
-- `GET /api/phone-endpoints/{id}/credentials` reveals the credentials again (`phoneEndpoint:update`; every call is audit-logged). `POST /api/phone-endpoints/{id}/credentials/rotate` mints a new password, returns the credentials once and resets the state to `initial`; the PBX's next REGISTER with the old password is refused. Both answer 404 for a client-mode registration.
+- `GET /api/phone-endpoints/{id}/credentials` reveals the credentials again (`phoneEndpoint:update`; every call is audit-logged). `POST /api/phone-endpoints/{id}/credentials` mints a new password, returns the credentials once and resets the state to `initial`; the PBX's next REGISTER with the old password is refused. Both answer 404 for a client-mode registration.
 - `GET /api/phone-endpoints/{id}/bindings` returns the mirror; `?live=1` asks the owning node through the node API (same proxying and status codes as the trace routes).
 
 Design record: `aplisay-strategy/implementation/regserver-tactical-spec.md`.
